ml/requirements.txt
# VOIZ Data Analysis API 의존성 라이브러리

# FastAPI 웹 프레임워크
fastapi>=0.104.1

# ASGI 서버 (FastAPI 실행용) 
uvicorn[standard]>=0.24.0

# 데이터 처리 및 수치 계산
numpy>=1.26.0

# 데이터 검증 및 설정 관리
pydantic>=2.5.0

# 타입 힌트 지원 (Python 3.8+ 에서는 기본 포함이지만 명시적으로 포함)
typing-extensions>=4.8.0

# 추가 유틸리티 (선택사항)
python-multipart==0.0.6

# OpenAI API 클라이언트
openai==1.54.3

# 환경변수 로딩
python-dotenv==1.0.0

# 추가 의존성
setuptools>=65.0.0 

# Google OCR API 클라이언트
google-cloud-vision==3.10.2

# 멀티파트 폼데이터 지원
python-multipart==0.0.20